pyproject.toml
[project]
name = "Orbi"
version = "0.1.0"
description = "Conversational travel agent"
requires-python = ">=3.10"

dependencies = [
    "langchain[anthropic,google-genai,openai]>=1.2.0",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=3.0.2",
    "langmem>=0.0.30",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]